+++ tests/testthat/test-comment.R
@@ -1,9 +1,8 @@
 context("comment")
 
 test_that("Comment creation works", {
-  x <- xml_new_document()
-  r <- xml_add_child(x, "root")
-  xml_add_child(r, xml_comment("Hello!"))
+  x <- xml_new_root("root")
+  xml_add_child(x, xml_comment("Hello!"))
   expect_identical(as.character(x),  "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<root><!--Hello!--></root>\n")
 })